requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.5.0
fastapi==0.104.1
uvicorn[standard]==0.24.0

# LangChain ecosystem (compatible versions)
langchain==0.1.0
langchain-openai==0.0.5
langgraph==0.0.15
langchain-community==0.0.10

# Utilities
typing-extensions==4.8.0
click==8.1.7
httpx==0.25.0
anyio==3.7.1